Update Long to Integer for all Ids

+++ src/main/java/com/management/person/repository/UserAccountRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Integer> {
 
     // Para login do Owner
     Optional<UserAccount> findByEmail(String email);
